+++ src/core/tamper/uninitializedvariable.py
@@ -13,6 +13,7 @@ the Free Software Foundation, either version 3 of the License, or
 For more see the file 'readme/COPYING' for copying permission.
 """
 
+import re
 import sys
 import random
 import string
@@ -20,7 +21,7 @@ import string
 from src.utils import settings
 
 """
-About: Adds uninitialized bash variables at the end of each command of the generated payloads (for *nix targets).
+About: Adds uninitialized bash variables between the characters of each command of the generated payloads.
 Notes: This tamper script works against *nix targets.
 Reference: https://www.secjuice.com/web-application-firewall-waf-evasion/
 """
@@ -33,7 +34,20 @@ if not settings.TAMPER_SCRIPTS[__tamper__]:
 def tamper(payload):
   def add_uninitialized_variable(payload):
     settings.TAMPER_SCRIPTS[__tamper__] = True
-    payload = payload.replace(settings.WHITESPACE[0], "$A" + settings.WHITESPACE[0])
+    rep = {
+            "${uv}I${uv}F${uv}S": "IFS",
+            "${uv}i${uv}f": "if", 
+            "${uv}t${uv}h${uv}e${uv}n": "then",
+            "${uv}e${uv}l${uv}s${uv}e": "else",
+            "${uv}f${uv}i": "fi",
+            "${uv}s${uv}t${uv}r": "str",
+            "${uv}c${uv}m${uv}d": "cmd",
+            "${uv}c${uv}ha${uv}r": "char"
+          }
+    payload = re.sub(r'([b-zD-Z])', r"${uv}\1", payload)
+    rep = dict((re.escape(k), v) for k, v in rep.items())
+    pattern = re.compile("|".join(rep.keys()))
+    payload = pattern.sub(lambda m: rep[re.escape(m.group(0))], payload)
     return payload
 
   if settings.TARGET_OS != "win":